refactor: move common to folder
