+++ client/routes/_authenticated/expenses/$expenseId/-common/ExpenseItemFieldGroup.tsx
@@ -1,10 +1,10 @@
 import { useMutation } from '@tanstack/react-query';
-import { withFieldGroup } from '../../../../components/Form';
-import { queryClient, trpc } from '../../../../trpc';
-import { defaultExpenseItem, type ExpenseFormData } from './-expense.common';
+import { withFieldGroup } from '../../../../../components/Form';
+import { queryClient, trpc } from '../../../../../trpc';
+import { defaultExpenseItem, type ExpenseFormData } from '.';
 import { X } from 'lucide-react';
 import type { DeepKeys, DeepValue } from '@tanstack/react-form';
-import { currencyNumberFormat } from '../../../../utils';
+import { currencyNumberFormat } from '../../../../../utils';
 
 type TGetFormField = <TField extends DeepKeys<ExpenseFormData>>(field: TField) => DeepValue<ExpenseFormData, TField>;
