+++ client/routes/_authenticated/expenses/$expenseId/geolocation.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import { setCurrentLocation, useExpenseForm } from './-expense.common';
+import { setCurrentLocation, useExpenseForm } from './-common';
 import { coordinateFormat, SG_CENTER } from '../../../../utils';
 import { APIProvider, ControlPosition, Map, Marker } from '@vis.gl/react-google-maps';
